add & modify commands

+++ powershell/Microsoft.PowerShell_profile.ps1
@@ -1,13 +1,3 @@
-# to be fancy
-neofetch
-
-# load oh my posh 
-oh-my-posh init pwsh --config "$env:USERPROFILE\clean-detailed.omp.json" | Invoke-Expression
-
-#34de4b3d-13a8-4540-b76d-b9e8d3851756 PowerToys CommandNotFound module
-Import-Module "$env:LOCALAPPDATA\PowerToys\WinUI3Apps\..\WinGetCommandNotFound.psd1"
-#34de4b3d-13a8-4540-b76d-b9e8d3851756
-
 # -------------------
 # SET COMMAND ALIASES AND STUFF
 
@@ -19,14 +9,20 @@ function psreload { pwsh }
 function psedit { micro $PROFILE }
 
 # alias for clearing the terminal
-Set-Alias -Name clear -Value cls
+Set-Alias -Name clear -Value Clear-Host
 
 # set history command to show full history
 function Show-History { Get-Content (Get-PSReadLineOption).HistorySavePath }
 Set-Alias -Name history -Value Show-History
 
+# find executable/script path
+function which($name) {
+	Get-Command $name | Select-Object -ExpandProperty Definition
+}
+
 # change network settings
 Set-Alias -Name chnet -Value "$env:USERPROFILE\chnet.bat"
+# function chnet { Start-Process "$env:USERPROFILE\chnet.bat" -Verb runAs -ArgumentList $arg }
 
 # alias for micro editor
 Set-Alias -Name m -Value micro
@@ -34,71 +30,62 @@ Set-Alias -Name mic -Value micro
 
 # system scan and repair
 function SystemScanAndRepair {
-	echo "Starting system scan and repair using SFC & DISM..."
-	echo "> SFC /scannow"
+	"Starting system scan and repair using SFC & DISM..."
+	"> SFC /scannow"
 	SFC /scannow
-	echo "> DISM /Online /Cleanup-Image /CheckHealth"
+	
+	"> DISM /Online /Cleanup-Image /CheckHealth"
 	DISM /Online /Cleanup-Image /CheckHealth
-	echo "> DISM /Online /Cleanup-Image /ScanHealth"
+
+	"> DISM /Online /Cleanup-Image /ScanHealth"
 	DISM /Online /Cleanup-Image /ScanHealth
-	echo "> DISM /Online /Cleanup-Image /RestoreHealth"
+
+	"> DISM /Online /Cleanup-Image /RestoreHealth"
 	DISM /Online /Cleanup-Image /RestoreHealth
 }
 Set-Alias -Name sysscan -Value SystemScanAndRepair
 Set-Alias -Name sysscanrepair -Value SystemScanAndRepair
 Set-Alias -Name syssr -Value SystemScanAndRepair
 
-# Simple function to start a new elevated process. If arguments are supplied then 
-# a single command is started with admin rights; if not then a new admin instance
-# of PowerShell is started.
-function admin {
-	if ($args.Count -gt 0) {   
-		#$argList = "& '" + $args + "'"
-		#Start-Process "$psHome\pwsh.exe" -Verb runAs -ArgumentList $argList
-		echo "Enter your command manually in the newly window"
-		#Start-Process wt -Verb runAs
-		wt -p "PowerShell (Admin)" # this is using terminal profile name
-	} else {
-		#Start-Process "$psHome\pwsh.exe" -Verb runAs
-		#Start-Process wt -Verb runAs 
-		wt -p "PowerShell (Admin)" # this is using terminal profile name
-	}
-}
-
 # get public ipv4 address
 function getpubip {
 	echo "Public IPv4 address:"
-	(Invoke-WebRequest http://ifconfig.me/ip).Content
+	(Invoke-WebRequest https://ifconfig.me/ip).Content
 }
 
 # find a file in current directory and subdirectories
-function find($name) {
-	Get-ChildItem -recurse -filter "*${name}*" -ErrorAction SilentlyContinue | ForEach-Object {
-		$place_path = $_.directory
-		Write-Output "${place_path}\${name}"
+function find($file, $mode) {
+	Get-ChildItem -recurse -filter "*${file}*" -ErrorAction SilentlyContinue | ForEach-Object {
+	if ($mode -ieq "d"){
+		# detailed version with more info
+		$_
+	} else {
+		# normal version, only filepaths
+		# if printed as string, only outputs full path, not all info
+		"$_"
+		}
 	}
 }
-# detailed version
-function dfind($name){
-	Get-ChildItem -recurse -filter "*${name}*" -ErrorAction SilentlyContinue | ForEach-Object {
-		echo $_
+
+function proc($action, $name){
+	if ($action -ieq "get"){
+		# get process by name
+		Get-Process $name
+	} elseif ($action -ieq "kill"){
+		# kill process by process name
+		Get-Process $name -ErrorAction SilentlyContinue | Stop-Process
+	} else {
+		Write-Host "invalid parameter for action: $action"
+		Write-Host "{get, kill}"
 	}
 }
 
-# kill process by process name
-function killproc($name) {
-	Get-Process $name -ErrorAction SilentlyContinue | Stop-Process
-}
-
-# get process by name
-function getproc($name){
-	Get-Process $name
-}
-
-# create file
+# create empty file
 function mkfile($fn){
-	"" >> $fn #if file exists, appends an empty string to its content
+	"" >> $fn  # if the file exists, appends an empty string to its content
 }
+Set-Alias -Name touch -Value mkfile
+
 function showwifi($name){
 	netsh wlan show profile name=$name key=clear
 }
@@ -121,3 +108,28 @@ function optvid ( $video, $outname, $quality = 'high' ){
 	ffmpeg -i $video -vcodec libx265 -crf $crf "$outname"
 }
 #--------------------
+# INIT STUFF
+
+# Print "Administrator" if the instance is elevated
+if ( ([Security.Principal.WindowsPrincipal] [Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole] "Administrator") ){
+	"Administrator"
+}
+
+# to be fancy
+neofetch
+
+# load oh my posh 
+oh-my-posh init pwsh --config "$env:USERPROFILE\clean-detailed.omp.json" | Invoke-Expression
+
+#34de4b3d-13a8-4540-b76d-b9e8d3851756 PowerToys CommandNotFound module
+Import-Module "$env:LOCALAPPDATA\PowerToys\WinUI3Apps\..\WinGetCommandNotFound.psd1"
+#34de4b3d-13a8-4540-b76d-b9e8d3851756
+
+# Import the PSReadLine module
+Import-Module PSReadLine
+
+# Enable predictive intellisense
+Set-PSReadLineOption -PredictionSource History
+# Set the UpArrow and DownArrow keys to search the command history
+Set-PSReadLineKeyHandler -Key UpArrow -Function HistorySearchBackward
+Set-PSReadLineKeyHandler -Key DownArrow -Function HistorySearchForward
